Add method to load game from RAM
This allows us to load games directly from a ByteArray, which is very
much ideal when bundling game files in the assets directory of the app.
Since assets cannot be accessed directly, they must be first copied out
of the app. We can avoid this by loading the game directly from a
ByteArray.

Signed-off-by: tytydraco <tylernij@gmail.com>